Likelihood: review 1- fix min needs cast to ulong for both params
for Win/AppVeynor

+++ src/nupic/algorithms/AnomalyLikelihood.cpp
@@ -471,7 +471,7 @@ vector<Real>  updateAnomalyLikelihoods(vector<Real> anomalyScores, UInt verbosit
   // Filter the likelihood values. First we prepend the historical likelihoods
   // to the current set. Then we filter the values.  We peel off the likelihoods
   // to return and the last windowSize values to store for later.
-  UInt toCrop = min((unsigned long)this->averagedAnomaly.getMaxSize(), runningLikelihoods.size());
+  UInt toCrop = min((unsigned long)this->averagedAnomaly.getMaxSize(), (unsigned long)runningLikelihoods.size());
   this->runningLikelihoods.insert(runningLikelihoods.end() - toCrop, likelihoods.begin(),likelihoods.end()); //append & crop
   assert(this->runningLikelihoods.size() <= this->averagedAnomaly.getMaxSize());
